LayoutTests/fast/js/object-extra-comma-expected.txt

This test checks some object construction cases, including 5939: final comma in javascript object prevents parsing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS var foo = { 'bar' : 'YES' }; foo.bar is 'YES'
PASS var foo = { 'bar' : 'YES', }; foo.bar is 'YES'
PASS var foo = { 'bar' : 'YES' , }; foo.bar is 'YES'
PASS var foo = { , 'bar' : 'YES' }; foo.bar threw exception SyntaxError: Unexpected token ','.
PASS var foo = { 'bar' : 'YES',, }; foo.bar threw exception SyntaxError: Unexpected token ','.
PASS successfullyParsed is true

TEST COMPLETE
